Can calculate lunar rising and setting
Needs more testing. Lunar numbers (esp. rising and setting vs. transit) aren't coming up exactly the same on the small number of test-cases, probably because the moon moves very fast across the sky and small input differences can mean minutes in difference in results.

Also, need to test the case where the moonrise or set is on a previous or following day.